requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic>=2.0
python-dotenv>=1.0
requests>=2.31
pyspark>=3.5,<3.6
boto3>=1.34
mysql-connector-python>=9.0
